moved weather into its own module
