+++ server.js
@@ -5,15 +5,11 @@ const app = express();
 const cors = require('cors');
 app.use(cors());
 const PORT = process.env.PORT || 5001;
-const weatherKey = process.env.WEATHER_API_KEY;
 const movieKey = process.env.MOVIE_API_KEY;
 const axios = require('axios').default;
 
-console.log('this is the port #', PORT);
-console.log('this is the weather key', weatherKey);
-console.log('this is the movie key', movieKey);
-
-app.get('/weather', handleWeather);
+const Weather = require('./weather');
+app.get('/weather', Weather.getForecast);
 app.get('/movies', handleMovie);
 
 function handleMovie(request, response) {
@@ -50,45 +46,7 @@ function handleMovie(request, response) {
 		});
 }
 
-function handleWeather(request, response) {
-	console.log('server side ', request.query);
-	let cityQuery = request.query.city;
-	let lonQuery = request.query.lon;
-	let latQuery = request.query.lat;
-	console.log(
-		`server has recieved the following parameters: ${cityQuery} + ${lonQuery} + ${latQuery}`
-	);
-	axios({
-		method: 'get',
-		url: 'http://api.weatherbit.io/v2.0/forecast/daily',
-		params: {
-			city: cityQuery,
-			lat: latQuery,
-			lon: lonQuery,
-			key: weatherKey,
-		},
-		transformResponse: [
-			function (data) {
-				// console.log('this is our data',  JSON.parse(data));
-				let betterData = JSON.parse(data);
-				let weatherArray = [];
-				for (let i = 0; i < 3; i++) {
-					let description = betterData.data[i].weather.description;
-					let dateTime = betterData.data[i].datetime;
-					weatherArray.push({ description: description, datetime: dateTime });
-				}
-				return weatherArray;
-			},
-		],
-	})
-		.then((res) => {
-			console.log(response.data);
-			response.send(res.data);
-		})
-		.catch((err) => {
-			console.error(err);
-		});
-}
+
 
 app.listen(PORT, () => {
 	console.log(`Server listening on ${PORT}`);
